requirements.txt
altgraph @ file:///AppleInternal/Library/BuildRoots/860631e9-c1c5-11ee-98ee-b6ef2fd8d87b/Library/Caches/com.apple.xbs/Sources/python3/altgraph-0.17.2-py2.py3-none-any.whl
attrs==25.3.0
aws-cdk-lib==2.188.0
aws-cdk.asset-awscli-v1==2.2.230
aws-cdk.asset-kubectl-v20==2.1.4
aws-cdk.asset-node-proxy-agent-v6==2.1.0
aws-cdk.cloud-assembly-schema==41.2.0
boto3==1.34.88
botocore==1.34.88
cattrs==24.1.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
constructs==10.4.2
cryptography==42.0.5
exceptiongroup==1.2.2
future @ file:///AppleInternal/Library/BuildRoots/860631e9-c1c5-11ee-98ee-b6ef2fd8d87b/Library/Caches/com.apple.xbs/Sources/python3/future-0.18.2-py3-none-any.whl
idna==3.7
importlib_resources==6.5.2
jmespath==1.0.1
jsii==1.111.0
macholib @ file:///AppleInternal/Library/BuildRoots/860631e9-c1c5-11ee-98ee-b6ef2fd8d87b/Library/Caches/com.apple.xbs/Sources/python3/macholib-1.15.2-py2.py3-none-any.whl
publication==0.0.3
pycparser==2.22
pycrypto==2.6.1
PyJWT==2.8.0
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.31.0
s3transfer==0.10.1
simple-crypt==4.1.7
six @ file:///AppleInternal/Library/BuildRoots/860631e9-c1c5-11ee-98ee-b6ef2fd8d87b/Library/Caches/com.apple.xbs/Sources/python3/six-1.15.0-py2.py3-none-any.whl
typeguard==2.13.3
typing_extensions==4.13.1
urllib3==1.26.18
zipp==3.21.0
